# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API - Companies Management",
    "version": "1.0.0",
    "description": "Documentación de la API para gestionar empresas y sus almacenes."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
    "/companies": {
      "get": {
        "summary": "Obtener todas las empresas",
        "tags": ["Companies"],
        "responses": {
          "200": {
            "description": "Lista de empresas obtenida correctamente"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "summary": "Crear una nueva empresa",
        "tags": ["Companies"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "OpenAI S.A." },
                  "address": { "type": "string", "example": "Av. 123 Calle 45" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Empresa creada correctamente" },
          "400": { "description": "Datos inválidos" }
        }
      }
    },
    "/companies/{id}": {
      "get": {
        "summary": "Obtener una empresa por ID",
        "tags": ["Companies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID de la empresa"
          }
        ],
        "responses": {
          "200": { "description": "Empresa encontrada" },
          "404": { "description": "Empresa no encontrada" }
        }
      },
      "put": {
        "summary": "Actualizar una empresa existente",
        "tags": ["Companies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID de la empresa"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "OpenAI Updated S.A." },
                  "address": { "type": "string", "example": "Nueva Calle 56 #78" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Empresa actualizada correctamente" },
          "400": { "description": "Datos inválidos" },
          "404": { "description": "Empresa no encontrada" }
        }
      },
      "delete": {
        "summary": "Eliminar una empresa por ID",
        "tags": ["Companies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID de la empresa"
          }
        ],
        "responses": {
          "200": { "description": "Empresa eliminada correctamente" },
          "404": { "description": "Empresa no encontrada" }
        }
      }
    },
    "/companies/{id}/warehouses": {
      "get": {
        "summary": "Obtener los almacenes de una empresa",
        "tags": ["Companies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID de la empresa"
          }
        ],
        "responses": {
          "200": { "description": "Lista de almacenes obtenida correctamente" },
          "404": { "description": "Empresa no encontrada" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Company": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "OpenAI S.A." },
          "address": { "type": "string", "example": "Av. 123 Calle 45" },
          "warehouses": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Warehouse" }
          }
        }
      },
      "Warehouse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 10 },
          "name": { "type": "string", "example": "Bodega Central" },
          "location": { "type": "string", "example": "Medellín" }
        }
      }
    }
  }
}